Rewrite shell.nix to use shellFor and callCabal2nix
The old `shell.nix` manually listed the dependencies required by the shell. Worse, this manual list was out of sync with the actual dependencies needed by `gerber` and `gerber-diagrams`; e.g., it listed `diagrams` as a dependency, which was failing because of an upper bound in one of `diagrams-contrib`'s dependencies, but `gerber-diagrams` doesn't even need `diagrams`, it only needs `diagrams-lib` and `diagrams-cairo`. Using `callCabal2nix` and `shellFor` will automatically generate a shell with the exact dependencies specified by in the cabal files.