Attempt to draw all valid Gerber images correctly (#6)
* Attempt to draw all valid Gerber images correctly

The library initially only had limited support for Gerber images, just
enough for what it was used for, but since we are using it for more now
we needed to improve support.

The following commands were not supported but are now.
  - Circular, rectangular and orbround flashes with holes.
  - Polygon flashes.
  - Macro flash commands.
  - Block apertures.
  - Rotate, mirror and scale commands.
  - A valid Gerber X1 parser should also parse Gerber X2 commands but are
    free to ignore them, so these are parsed now but ignored when
    drawing.

To support flashes with holes and macros which flash as a single
polarity, but which are composed of additive and subtractive primitives,
the 'Diagrams.TwoD.Path.Boolean' module is used which offers limited
support for doing boolean operations on paths. The result of the
operations are not always correct but for most Gerbers we would
encounter in the wild it does a good enough job. A future improvement
could be to switch away from using this module or fixing the
deficiencies in `cubicbezier` package which is what it ultimately uses.

Golden tests were added for parsing sample files obtained from Ucamco's
website https://www.ucamco.com/en/gerber/downloads.
https://www.ucamco.com/files/downloads/file_en/423/gerber-file-format-test-files_en.zip?953d9c5162cda74a96615ed657afc8b7
The images defined by these can be compared with the reference viewer
here https://gerber-viewer.ucamco.com/index.html. Extra sample Gerbers
from the wild were also added to the set of golden tests.

All the non reference samples parse and display correctly and almost all
the reference samples display correctly.

The reference samples `sample_macro.gbr` and `sample_macro_X1.gbr`
display mostly correctly except for the one primitive macro which is a
manually defined thermal. This is not displaying correctly due to the
limited ability of the 'Diagrams.TwoD.Path.Boolean' module, subtracting
wedges from a circle results in the circle disappearing and not gaps in
the circle.

Some files we encounter in the wild are invalid gerber files specifying
some unsupported macro primitives so instead of failing to parse we
parse them noting that they are invalid definitions.