Change memory layout of EncodedDecimal
The `EncodedDecimal` type was correct for representing encoded decimals
in the Gerber standard but it was a terribly wasteful way to store
numbers which are used as coordinates and which dominate a Gerber file.
For processing large Gerber files (over a 100 MB) it lead to a memory
explosion.

To fix this we keep the interface of the `EncodedDecimal` constructor
but we change the representation to more efficient packed number format.
All the tests still pass without modification.

In another branch I added binary serialization to the Gerber command and
before the change a 100 MB Gerber file would be serialized as a 500 MB
binary file but after the size is basically the same.